Attempt to fix Windows warning
I don't have access to Windows right now, so it's possible this won't work.

+++ src/Common/Utilities/src/TextObjectFormatter.cpp
@@ -143,7 +143,7 @@ namespace BitFunnel
     }
 
 
-    void TextObjectFormatter::Format(unsigned long value)
+    void TextObjectFormatter::Format(size_t value)
     {
         m_output << value;
     }
